Extract room Manger code
